evals/requirements.txt
mypy>=1.0.0
types-setuptools
openai>=1.0.0
prettytable>=3.0.0
pandas>=2.0.0
matplotlib>=3.0.0
ipython
pydantic
mitosheet
snowflake-sqlalchemy